Stage 4.8.5a: make the Linux oracle operational
Two bugs were keeping the engine from running on GCC/Clang Linux
even after the warning sweep was clean.

1. Char signedness mismatch — the bigger of the two.
   Voxlap's RLE walker at scum2line is `while (v[0]) v += v[0]*4;`
   where `v[0]` is the next-slab byte offset divided by 4. Wide
   carves (e.g. the oracle's 184-tall playable box) push that
   header past 127. With signed char the read sign-extends to
   negative and the loop walks backward through vbuf forever; what
   we'd been calling "the Linux startup hang" was a slow infinite
   loop in compilerle / scum2line with no infinite-loop watchdog
   in sight.
   The original Voxlap project assumed unsigned char throughout
   and built MSVC with `/J` (the vcxproj already had it). Add the
   GCC/Clang equivalent `-funsigned-char` to the non-MSVC branch
   of CMakeLists.txt.

2. getcube returns a pointer through int32_t.
   getcube's return-value contract is "0 = air, 1 = unexposed
   solid, else vbuf-column pointer cast to int". On LP64 the int32
   carrier truncates the upper 32 bits of the pointer; meltsphere /
   meltspans / meltfall then deref the truncated value via
   `*(int32_t *)i` and SIGSEGV.
   Widen getcube's return type to intptr_t (header + definition).
   Bit-identical on Windows MSVC x86 (intptr_t == int32_t).
   Existing callers that test `getcube() & ~1` for non-air-non-
   unexposed still work since pointer-aligned values clear bit 0.
   The buggy `setbrightnessfunc` in exports.c (line 34) which
   `& 0xffffff`-masks getcube's return as if it were a colour
   word remains broken — pre-existing, not exercised by the oracle.

After both fixes the Linux GCC oracle runs cleanly, exits 0, and
produces a full hashes.txt. 10 of 12 hashes match the Windows MSVC
goldens bit-for-bit; the two `sprite_*` hashes differ as predicted
in PORTING-X64.md (Stage 4.7.5/4.7.6 replaced rsqrtss with
1.0/sqrtf, which is more accurate but not bit-identical). Per-
platform golden tables come in 4.8.5b — this commit just makes
the engine RUN.

+++ voxlap/voxlap5.c
@@ -933,7 +933,7 @@ int32_t getfloorz (int32_t x, int32_t y, int32_t z)
 	//   0: air
 	//   1: unexposed solid
 	//else: address to color in vbuf (this can never be 0 or 1)
-int32_t getcube (int32_t x, int32_t y, int32_t z)
+intptr_t getcube (int32_t x, int32_t y, int32_t z)
 {
 	int32_t ceilnum;
 	char *v;
